修复singbox启动逻辑问题

+++ singbox/start.sh
@@ -92,11 +92,13 @@ fi
 
 #如果只有hy2和vmess+ws
 if [[ "$type" == "1" || "$type" == "3" || "$type" == "1.2" || "$type" == "2" || "$type" == "3.2" ]]; then
-    r=$(ps aux | grep cloudflare | grep -v grep | awk '{print $2}')
-    if [ -n "$r" ]; then
-      echo $r
-      kill -9 $r
-    fi
+   if [[  "$type" == "1.2" || "$type" == "2" || "$type" == "3.2" ]]; then
+      r=$(ps aux | grep cloudflare | grep -v grep | awk '{print $2}')
+      if [ -n "$r" ]; then
+        echo $r
+        kill -9 $r
+      fi
+   fi
     chmod +x ./serv00sb
     if ! ps aux | grep serv00sb | grep -v "grep" >/dev/null; then
       nohup ./serv00sb run -c ./config.json >/dev/null 2>&1 &
